+++ R/basis_set.R
@@ -100,113 +100,48 @@ stackplot.basis_set <- function(x, ...) {
 #' @return basis object.
 #' @export
 read_basis <- function(basis_file, ref = def_ref(), sort_basis = TRUE) {
+  
+  # open the file
   con  <- file(basis_file, open = "r")
   names <- vector()
   data <- vector()
   
   while (length(line <- readLines(con, n = 1, warn = FALSE)) > 0) {
+    # reading main header
     if (startsWith(line, " NDATAB = ")) {
       N <- as.integer(strsplit(trimws(line), "\\s+")[[1]][3])
-      data_lines <- ceiling(2 * N / 6) # assume 6 cols
     } else if (startsWith(line, " HZPPPM = ")) {
       bas_ft <- strsplit(trimws(line), "\\s+")[[1]][3]
-      bas_ft <- as.double(gsub(",", "", bas_ft))*1e6
+      bas_ft <- as.double(gsub(",", "", bas_ft)) * 1e6
     } else if (startsWith(line, " BADELT = ")) {
       bas_fs <- strsplit(trimws(line), "\\s+")[[1]][3]
-      bas_fs <- 1 / as.double(gsub(",", "",bas_fs))
-    } else if (endsWith(line, "$BASIS")) {
+      bas_fs <- 1 / as.double(gsub(",", "", bas_fs))
+    } else if (startsWith(line, " FMTBAS = ")) {
+      format <- (strsplit(trimws(line), " ")[[1]][3])
+      format <- gsub(",", "", format)
+      format <- gsub("'", "", format)
+      format <- gsub("\\(", "", format)
+      format <- gsub("\\)", "", format)
+    } else if (endsWith(line, "$BASIS")) { # we're at the end of the main header
+      cols <- as.numeric(substr(format, 1, 1))
+      data_lines <- ceiling(2 * N / cols)
       while (length(line <- readLines(con, n = 1, warn = FALSE)) > 0) {
         if (startsWith(line, " ID = ")) {
+          # reading metabolite header
           id <- (strsplit(trimws(line), " ")[[1]][3])
-          id <- gsub(",", "",id)
-          id <- gsub("'", "",id)
+          id <- gsub(",", "", id)
+          id <- gsub("'", "", id)
           names <- c(names, id)
         } else if (endsWith(line, "$END")) {
-          x <- utils::read.fortran(con, "6F13.0", n = data_lines)
+          # read data points
+          x <- utils::read.table(con, nrows = data_lines, fill = TRUE)
           data_pts <- as.vector(t(as.matrix(x)))
           data_pts <- data_pts[seq(1, 2 * N, 2)] +
                       1i * data_pts[seq(2, 2 * N, 2)]
-          data_pts <- pracma::ifftshift(data_pts)
-          data <- cbind(data, data_pts)
-          break
-        }
-      }
-    }
-  }
-  close(con)
-  
-  basis_set <- list(data = data, N = N, fs = bas_fs, ft = bas_ft, 
-                    names = names, ref = ref)
-  
-  class(basis_set) <- "basis_set"
-  
-  if (sort_basis) basis_set <- sort_basis(basis_set)
-  
-  dimnames(basis_set$data) <- NULL
-
-  return(basis_set)
-}
-
-#' Read a basis file in LCModel .basis format (for testing only).
-#' @param basis_file path to basis file.
-#' @param ref assumed ppm reference value.
-#' @param sort_basis sort the basis set based on signal names.
-#' @return basis object.
-#' @export
-read_basis_ac <- function(basis_file, ref = def_ref(), sort_basis = TRUE) {
-  con  <- file(basis_file, open = "r")
-  names <- vector()
-  data <- vector()
-  
-  while (length(line <- readLines(con, n = 1, warn = FALSE)) > 0) {
-    
-    # ARC 2021-07-15 : handle varying parameter spacing and column layout 
-    line<-trimws(line);
-    # split by "=", surrounded by arbitrary whitespace.
-    # some lines have a "," suffix; we can cut this out in the same operation
-    tokens <- strsplit(line, "[[:space:]]*[=,][[:space:]]*")[[1]]
-    if (is.na(tokens[1])) {
-      # empty line
-      next;
-    }         
-    param <- toupper(tokens[1])
-    if (param=="NDATAB") {
-      N <- as.integer(tokens[2])
-    } else if (param=="HZPPPM") {
-      bas_ft <- as.double(tokens[2])*1e6
-    } else if (param=="BADELT") {
-      bas_fs <- 1/as.double(tokens[2])
-    } else if (endsWith(line, "$BASIS")) {
-      while (length(line <- readLines(con, n = 1, warn = FALSE)) > 0) {
-        line <- trimws(line);
-        tokens <- strsplit(line, "[[:space:]]*[=,][[:space:]]*")[[1]]
-        param <- toupper(tokens[1]);
-        if (is.na(tokens[1])) {
-          next;
-        }
-        if (param=="METABO") { # formerly ID
-          id <- trimws(gsub("'", "",tokens[2]))
-          names <- c(names, id)
-        } else if (endsWith(line, "$END")) {
-          # ARC 2021-07-15: auto-detect column layout {{{
-          # read a single line, then rewind
-          fp <- seek(con,origin='current')
-          l1 <- readLines(con,n=1,warn=FALSE); 
-          fpn <- seek(con,origin='start',where=fp);
           
-          tokens <- strsplit(trimws(l1),"[[:space:]]+")[[1]];
-          cols <- length(tokens);
-          width <- ceiling(nchar(l1)/cols);
-          fmt <- sprintf("%dF%d.0",cols,width);
-          # }}}
-          
-          data_lines <- ceiling(2 * N / cols)
-          x <- utils::read.fortran(con, fmt, n = data_lines)
-          data_pts <- as.vector(t(as.matrix(x)))
-          data_pts <- data_pts[seq(1, 2 * N, 2)] +
-            1i * data_pts[seq(2, 2 * N, 2)]
+          # transform to time-domain
           data_pts <- pracma::ifftshift(data_pts)
-          data <- cbind(data, data_pts)
+          data     <- cbind(data, data_pts) # slow?
           break
         }
       }