+++ NEWS.md
@@ -1,6 +1,7 @@
 # spant 2.5.0
 * Improved function documentation for water concentration scaling.
 * Added water and tCr scaling examples to the introduction vignette.
+* Better error checking for reading ima files.
 
 # spant 2.4.0
 * Performance improvement for reading VD format TWIX files.