+++ NEWS.md
@@ -3,6 +3,8 @@
 * Added water and tCr scaling examples to the introduction vignette.
 * Better error checking for reading ima files.
 * read_basis function improved to work with variable formatting.
+* Added helper functions for working with basis set objects: 
+scale_basis_from_singlet and get_basis_subset.
 
 # spant 2.4.0
 * Performance improvement for reading VD format TWIX files.