+++ NEWS.md
@@ -5,6 +5,8 @@
 * read_basis function improved to work with variable formatting.
 * Added helper functions for working with basis set objects: 
 scale_basis_from_singlet, resample_basis and get_basis_subset.
+* Bug fix for STEAM basis simulation being incorrectly scaled for internal
+water scaling methods.
 
 # spant 2.4.0
 * Performance improvement for reading VD format TWIX files.